requirements.txt
anyascii==0.3.1
asgiref==3.5.2
beautifulsoup4==4.9.3
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==3.2.15
django-allauth==0.51.0
django-cors-headers==3.13.0
django-filter==2.4.0
django-modelcluster==5.3
django-ninja==0.19.1
django-taggit==1.5.1
django-treebeard==4.5.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
html5lib==1.1
idna==3.3
l18n==2021.3
oauthlib==3.2.0
openpyxl==3.0.10
Pillow==8.4.0
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tablib==3.2.1
telepath==0.2
typing_extensions==4.3.0
urllib3==1.26.11
wagtail==2.14.2
webencodings==0.5.1
Willow==1.4.1
xlrd==2.0.1
XlsxWriter==1.4.5
xlwt==1.3.0
